+++ examples/scheduling-mime-email-with-freemarker-example/src/main/java/com/test/MimeEmailWithFreemarkerSchedulingApplication.java
@@ -12,6 +12,7 @@ import org.springframework.context.ConfigurableApplicationContext;
 
 import javax.annotation.PostConstruct;
 import java.io.UnsupportedEncodingException;
+import java.net.URISyntaxException;
 import java.util.Timer;
 import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
@@ -35,7 +36,7 @@ public class MimeEmailWithFreemarkerSchedulingApplication implements Application
     }
 
     @PostConstruct
-    public void sendEmail() throws UnsupportedEncodingException, InterruptedException, CannotSendEmailException {
+    public void sendEmail() throws UnsupportedEncodingException, InterruptedException, CannotSendEmailException, URISyntaxException {
         testService.scheduleSixMimeEmails();
 
         close();